cleaned up the up60p_restore_beast_main.c file with headers and new fresh .c files in the program to be more organized and compact
